+++ src/data-feeds/MockOffchainAggregator.sol
@@ -2,6 +2,9 @@
 pragma solidity ^0.8.0;
 
 contract MockOffchainAggregator {
+    int192 private constant MIN_ANSWER_POSSIBLE = 1;
+    int192 private constant MAX_ANSWER_POSSIBLE = 95780971304118053647396689196894323976171195136475135; // type(uint176).max
+
     uint8 public decimals;
     int256 public latestAnswer;
     uint256 public latestTimestamp;
@@ -23,8 +26,8 @@ contract MockOffchainAggregator {
         updateAnswer(_initialAnswer);
         // If the minAnswer has a value of 1 and the maxAnswer has a value of 95780971304118053647396689196894323976171195136475135 then that theoretically means there is no min or max for that feed
         // If the minAnswer and maxAnswer values are set to other than those mentioned above, then there are actually min and max for that feed - which you will need to normalize using the demicals value
-        minAnswer = 1;
-        maxAnswer = type(int192).max;
+        minAnswer = MIN_ANSWER_POSSIBLE;
+        maxAnswer = MAX_ANSWER_POSSIBLE;
     }
 
     function updateAnswer(int256 _answer) public {
@@ -68,6 +71,10 @@ contract MockOffchainAggregator {
     }
 
     function updateMinAndMaxAnswers(int192 _minAnswer, int192 _maxAnswer) external {
+        require(_minAnswer < _maxAnswer, "minAnswer must be less than maxAnswer");
+        require(_minAnswer >= MIN_ANSWER_POSSIBLE, "minAnswer is too low");
+        require(_maxAnswer <= MAX_ANSWER_POSSIBLE, "maxAnswer is too high");
+
         minAnswer = _minAnswer;
         maxAnswer = _maxAnswer;
     }